admin page wrap up
